# focaly-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Focaly API
  description: |
    Backend API for the Focaly Study Management Mobile App.
    See `specs/001-focaly-backend/spec.md` for user-facing requirements and
    `specs/001-focaly-backend/data-model.md` for entity shapes.

    All endpoints versioned under `/v1`. Bearer auth (access credential) is the
    default global security requirement; `Public` endpoints opt out and are
    annotated below. Refresh uses a separate `bearerRefresh` scheme.

    Error envelope (shared component `ErrorResponse`):
    `{ code: string, message: string, details?: object }`
  version: '1.0.0'
servers:
  - url: https://api.focaly.app/v1
    description: production
  - url: https://staging-api.focaly.app/v1
    description: staging
  - url: http://localhost:3000/v1
    description: local dev
security:
  - bearerAccess: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile, plan, settings
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/User' } } },
          }
    patch:
      tags: [Users]
      summary: Update profile (name, locale, timezone, avatar URL)
      requestBody: { $ref: '#/components/requestBodies/UpdateUserBody' }
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/User' } } },
          }
    delete:
      tags: [Users]
      summary: Soft-delete the account; hard-purge within 30 days (FR-009, SC-012)
      responses: { '204': { description: No Content } }
  /users/me/settings:
    patch:
      tags: [Users]
      summary: Update notification preferences and focusMode
      requestBody: { $ref: '#/components/requestBodies/UpdateSettingsBody' }
      responses:
        '200':
          {
            description: OK,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/UserSettings' } } },
          }
  /users/me/avatar:
    post:
      tags: [Users]
      summary: Upload a new avatar (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          {
            description: OK,
            content:
              {
                application/json:
                  {
                    schema:
                      { type: object, properties: { avatarUrl: { type: string, format: uri } } },
                  },
              },
          }
  /users/me/fcm-token:
    post:
      tags: [Users]
      summary: Register / update the push token for the current device session
      requestBody: { $ref: '#/components/requestBodies/FcmTokenBody' }
      responses: { '204': { description: No Content } }

  # ────────────────────────────── Subjects ──────────────────────────────
...
  /analytics/summary:
    get:
      tags: [Analytics]
      summary: Totals over a date range (Premium for ranges beyond current week)
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date } }
        - { in: query, name: to, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          {
            description: OK,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AnalyticsSummary' } } },
          }
        '403': { $ref: '#/components/responses/PremiumRequired' }
  /analytics/by-subject:
    get:
      tags: [Analytics]
      summary: Minutes per subject in range (Premium)
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date } }
        - { in: query, name: to, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          {
            description: OK,
            content:
              {
                application/json:
                  {
                    schema: { type: array, items: { $ref: '#/components/schemas/PerSubjectStat' } },
                  },
              },
          }
        '403': { $ref: '#/components/responses/PremiumRequired' }
  /analytics/heatmap:
    get:
      tags: [Analytics]
      summary: Year-long daily focus-minutes heatmap (Premium)
      parameters: [{ in: query, name: year, required: true, schema: { type: integer } }]
      responses:
        '200':
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { type: array, items: { $ref: '#/components/schemas/HeatmapPoint' } } },
              },
          }
        '403': { $ref: '#/components/responses/PremiumRequired' }
  /analytics/performance:
    get:
      tags: [Analytics]
      summary: Completion + retention performance (Premium)
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date } }
        - { in: query, name: to, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          {
            description: OK,
            content:
              {
                application/json: { schema: { $ref: '#/components/schemas/AnalyticsPerformance' } },
              },
          }
        '403': { $ref: '#/components/responses/PremiumRequired' }

  # ──────────────────────────────── Uploads ─────────────────────────────
...
components:
  securitySchemes:
    bearerAccess:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-min access credential (FR-005)
    bearerRefresh:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 30-day rotating refresh credential (FR-005, FR-007)
  responses:
...
    PremiumRequired:
      description: Premium-only resource (FR-035)
      content:
        {
          application/json:
            {
              schema: { $ref: '#/components/schemas/ErrorResponse' },
              example:
                { code: PREMIUM_REQUIRED, message: 'Upgrade to premium to access this feature.' },
            },
        }
...
  requestBodies:
...
    UpdateUserBody:
      required: true
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties:
                    {
                      name: { type: string },
                      locale: { type: string },
                      timezone: { type: string },
                      avatarUrl: { type: string, format: uri },
                    },
                },
            },
        }
    UpdateSettingsBody:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/UserSettings' } } }
    FcmTokenBody:
      required: true
      content:
        {
          application/json:
            {
              schema:
                { type: object, required: [fcmToken], properties: { fcmToken: { type: string } } },
            },
        }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    UserSettings:
      type: object
      properties:
        locale: { type: string, example: en-US }
        timezone: { type: string, example: Europe/Berlin }
        focusMode: { type: boolean }
        notifications:
          type: object
          properties:
            reminders: { type: boolean }
            streak: { type: boolean }
            marketing: { type: boolean }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        avatarUrl: { type: string, format: uri }
        emailVerified: { type: boolean }
        plan: { type: string, enum: [free, premium] }
        premiumUntil: { type: string, format: date-time }
        settings: { $ref: '#/components/schemas/UserSettings' }
        totalPoints: { type: integer }
...
    AnalyticsSummary:
      type: object
      properties:
        range:
          {
            type: object,
            properties:
              { from: { type: string, format: date }, to: { type: string, format: date } },
          }
        totals:
          type: object
          properties:
            minutes: { type: integer }
            sessions: { type: integer }
            tasksCompleted: { type: integer }
        series:
          type: array
          items:
            {
              type: object,
              properties:
                {
                  date: { type: string, format: date },
                  minutes: { type: integer },
                  tasks: { type: integer },
                },
            }
    PerSubjectStat:
      type: object
      properties:
        subjectId: { type: string }
        name: { type: string }
        minutes: { type: integer }
        sessions: { type: integer }
    HeatmapPoint:
      type: object
      properties:
        date: { type: string, format: date }
        minutes: { type: integer }
    AnalyticsPerformance:
      type: object
      properties:
        completionRate: { type: number }
        streakRetention: { type: number }